pyproject.toml
# Copyright (C) Free Software Foundation, Inc. All rights reserved.
# Licensed under the AGPL-3.0-only License. See LICENSE in the project root
# for license information.

[build-system]
requires = ['setuptools>=61.0', 'setuptools-git-versioning<2']
build-backend = 'setuptools.build_meta'

[project]
name = 'XRoadServiceCatalogCollector'
description = 'X-Road Service Catalog Collector'
readme = 'README.md'
requires-python = '>=3.8'
license = {text = 'AGPL-3.0-only'}
keywords = ['X-Road', 'Service catalog collector']
authors = [{name = 'Pipin Fitriadi'}, {email = 'pipinfitriadi@gmail.com'}]
maintainers = [{name = 'Pipin Fitriadi'}, {email = 'pipinfitriadi@gmail.com'}]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Programming Language :: Python :: 3.8',
    'License :: OSI Approved :: GNU Affero General Public License v3',
    'Operating System :: OS Independent',
    'Framework :: FastAPI',
    'Topic :: File Formats :: JSON',
    'Topic :: Internet :: WWW/HTTP :: HTTP Servers',
    'Topic :: Text Processing',
    'Topic :: Utilities'
]
dependencies = ['python-dotenv', 'typer[all]']
dynamic = ['version']

[project.optional-dependencies]
restful-api = ['fastapi', 'uvicorn[standard]']

[project.urls]
Repository = 'https://github.com/jabardigitalservice/x-road-service-catalog-collector'

[tool.coverage.run]
relative_files = true

[tool.setuptools-git-versioning]
enabled = true
template = '{tag}'
dev_template = '{tag}'
dirty_template = '{tag}'
